HomeViewModel creation & HomeView refactor
