+++ FitnessApp/Home/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Zouz Bekdache on 3/1/2025.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
